+++ CHANGES.rst
@@ -8,6 +8,10 @@ Unreleased
 
 - Switch from unmaintained ``appdirs`` to the replacement ``platformdirs``.
 
+**Fixed**
+
+- An issue where the summary line of docstring was incorrectly wrapped.
+
 1.4.1 (2021/09/10)
 ------------------